% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{add_logo}
\alias{add_logo}
\title{Add a logo to a plot}
\usage{
add_logo(plot_path, logo_path, logo_position, logo_scale = 10)
}
\arguments{
\item{plot_path}{The file path of the plot image.}

\item{logo_path}{The file path of the logo image.}

\item{logo_position}{The position for the logo: "top right", "top left", "bottom right", or "bottom left".}

\item{logo_scale}{The scale factor for the logo size relative to the plot (default is 10).}
}
\value{
The plot image with the logo added.
}
\description{
This function adds a logo to a plot at a specified position and scale.
}
