% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_colors.R
\name{br_pal}
\alias{br_pal}
\title{A personal color palette to use with ggplot}
\usage{
br_pal()
}
\description{
A personal color palette to use with ggplot
}
\examples{
library(scales)
scales::show_col(br_pal()(9))
}
