% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fonts.R
\name{check_font}
\alias{check_font}
\title{Function to check font availability}
\usage{
check_font(font_name)
}
\arguments{
\item{font_name}{Name of the font to check}
}
\value{
A warning message if the font is not found.
}
\description{
Function to check font availability
}
