% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_emw_dark.R
\name{dark_geom_defaults}
\alias{dark_geom_defaults}
\title{Change geom defaults from black to custom lights for the Dark theme}
\usage{
dark_geom_defaults()
}
\description{
Change geom defaults from black to custom lights for the Dark theme
}
