% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_emw_dark.R
\docType{data}
\name{dark_pal}
\alias{dark_pal}
\title{Custom dark EMW palette}
\format{
An object of class \code{character} of length 9.
}
\usage{
dark_pal
}
\value{
A vector of dark EMW colors
}
\description{
Custom dark EMW palette
}
\keyword{datasets}
