% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_emw_dark.R
\name{dark_palette}
\alias{dark_palette}
\title{Custom dark EMW palette function}
\usage{
dark_palette(num_cols, var_type = c("discrete", "continuous"))
}
\arguments{
\item{num_cols}{Number of colors to return.}

\item{var_type}{Type of palette (discrete or continuous).}
}
\value{
A function that generates the EMW palette.
}
\description{
Custom dark EMW palette function
}
