% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fonts.R
\name{.pkg_fonts}
\alias{.pkg_fonts}
\title{Fonts included in the emwthemes package}
\usage{
.pkg_fonts()
}
\value{
A named list of font families and font styles
}
\description{
Fonts included in the emwthemes package
}
\examples{
\dontrun{
.pkg_fonts()
}
}
