% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_emw.R
\name{emw_geom_defaults}
\alias{emw_geom_defaults}
\title{Change geom defaults from black to custom lights for the Dark theme}
\usage{
emw_geom_defaults()
}
\description{
Change geom defaults from black to custom lights for the Dark theme
}
