% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_emw.R
\docType{data}
\name{emw_pal}
\alias{emw_pal}
\title{Custom theme palette}
\format{
An object of class \code{character} of length 9.
}
\usage{
emw_pal
}
\value{
A vector of personal theme colors
}
\description{
Custom theme palette
}
\keyword{datasets}
