% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emwColsPalettes.R
\docType{data}
\name{emwColsPalettes}
\alias{emwColsPalettes}
\title{Complete list of available palettes.}
\format{
An object of class \code{list} of length 37.
}
\usage{
emwColsPalettes
}
\description{
Use names(emwColsPalettes) to return all palette names.
}
\keyword{datasets}
