% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_font.R
\docType{data}
\name{font_urls}
\alias{font_urls}
\title{URL links to Google Fonts}
\format{
An object of class \code{data.frame} with 3 rows and 2 columns.
}
\usage{
font_urls
}
\value{
A data frame with font names and their respective URL links.
}
\description{
URL links to Google Fonts
}
\keyword{datasets}
