% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_colors.R
\name{ft_pal}
\alias{ft_pal}
\title{A personal color palette to use with ggplot}
\usage{
ft_pal()
}
\description{
A personal color palette to use with ggplot
}
\examples{
library(scales)
scales::show_col(ft_pal()(9))
}
