% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpme.R
\name{get_emw_image}
\alias{get_emw_image}
\title{Commonly used images}
\usage{
get_emw_image(folder_path, img_name)
}
\arguments{
\item{folder_path}{Names of addition path required to locate items within runza folder after \verb{https://emmarshall.github.io/runza/img/...}}

\item{img_name}{Name of column to locate individual item within folder}
}
\value{
Character vector of URLs
}
\description{
Returns the URL for the image based on \code{folder_path} and \code{name}
}
\examples{
library(dplyr)
## Need to set folder path correctly each time
emwthemes::hindsight |>
  mutate(
  ## Set folder path
  image = get_emw_image("hind/", timing))
}
