% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fonts.R
\name{import_daughter_font}
\alias{import_daughter_font}
\alias{import_shadows_font}
\alias{import_monolisa_font}
\alias{import_dmserif_font}
\alias{import_dmsans_font}
\alias{import_metropolis_font}
\alias{import_oswald_font}
\title{Import "Architects Daughter" font}
\usage{
import_daughter_font()

import_shadows_font()

import_monolisa_font()

import_dmserif_font()

import_dmsans_font()

import_metropolis_font()

import_oswald_font()
}
\description{
Import "Architects Daughter" font

Import "Shadows Into Light Two" font

Import "Monolisaemw" font

Import "DM Serif" font

Import "DM Sans" font

Import "Metropolis" font

Import "Oswald" font
}
\details{
import_*() functions taken from hrbrthemes and Jon Bernhardt..
Created by
You may still need to install each font on your system directly by finding the .ttf file and clicking "Install".

import_*() functions taken from hrbrthemes.
You may still need to install each font on your system directly by finding the .ttf file and clicking "Install".

import_*() functions taken from hrbrthemes and tvthemes
You may still need to install each font on your system directly by finding the .ttf file and clicking "Install".

import_*() functions taken from hrbrthemes.
You may still need to install each font on your system directly by finding the .ttf file and clicking "Install".

import_*() functions taken from hrbrthemes.
You may still need to install each font on your system directly by finding the .ttf file and clicking "Install".

import_*() functions taken from hrbrthemes.
You may still need to install each font on your system directly by finding the .ttf file and clicking "Install".

import_*() functions taken from hrbrthemes.
You may still need to install each font on your system directly by finding the .ttf file and clicking "Install".
}
\seealso{
\code{\link[extrafont]{font_import}}

\code{\link[extrafont]{font_import}}

\code{\link[extrafont]{font_import}}

\code{\link[extrafont]{font_import}}

\code{\link[extrafont]{font_import}}

\code{\link[extrafont]{font_import}}

\code{\link[extrafont]{font_import}}
}
