% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_img.R
\name{link_to_img}
\alias{link_to_img}
\title{Create an HTML image tag with a specified width taken from Thomas Mock blog post}
\usage{
link_to_img(x, width)
}
\arguments{
\item{x}{The image source URL.}

\item{width}{The width of the image in pixels.}
}
\value{
An HTML image tag.
}
\description{
Create an HTML image tag with a specified width taken from Thomas Mock blog post
}
