% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_emw.R
\name{scale_color_emw_d}
\alias{scale_color_emw_d}
\title{Plotting with personal palettes for color ggplot2 - inspired by emwCols()}
\usage{
scale_color_emw_d(name, direction = 1, legend_title = NULL, ...)
}
\arguments{
\item{name}{Name of Palette. Run \code{names(emwColsPalettes)} to view options.}

\item{direction}{Sets order of colors. Default palette is 1. If direction is -1,
palette color order is reversed}

\item{legend_title}{Character string specifying legend title. Default \code{NULL}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{scale_color_gradientn}}}
}
\value{
A ggplot object defining a continuous color scale for use with ggplot2.
}
\description{
Plotting with personal palettes for color ggplot2 - inspired by emwCols()
}
\examples{
library(ggplot2)
ggplot(data=mtcars, aes(x=mpg, y=disp, color=wt)) +
  geom_point() +
  scale_color_emw_c("Reds")
}
