% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_emw.R
\name{scale_fill_emw_c}
\alias{scale_fill_emw_c}
\title{Plotting with emwCols palettes for color ggplot2}
\usage{
scale_fill_emw_c(name, direction = 1, legend_title = NULL, ...)
}
\arguments{
\item{name}{Name of Palette. Run \code{names(emwColsPalettes)} to view options.}

\item{direction}{Sets order of colors. Default palette is 1. If direction is -1,
palette color order is reversed}

\item{legend_title}{Character string specifying legend title. Default \code{NULL}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{scale_fill_gradientn}}}
}
\value{
A ggproto object defining a continuous fill scale for use with ggplot2.
}
\description{
Plotting with emwCols palettes for color ggplot2
}
\examples{
library(ggplot2)
ggplot(data=mtcars, aes(x=mpg, y=disp, fill=wt)) +
  geom_point(pch = 21) +
  scale_fill_emw_c("Corals")
}
