% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_defaults.R
\name{scale_x_percent}
\alias{scale_x_percent}
\alias{scale_y_percent}
\alias{scale_x_comma}
\alias{scale_y_comma}
\title{X & Y scales with opinionated pre-sets for percent & comma label formats}
\usage{
scale_x_percent(
  name = NULL,
  breaks = NULL,
  minor_breaks = NULL,
  labels = NULL,
  limits = NULL,
  expand = c(0.01, 0),
  oob = scales::censor,
  na.value = NA_real_,
  trans = "identity",
  position = "bottom",
  ...
)

scale_y_percent(
  name = NULL,
  breaks = NULL,
  minor_breaks = NULL,
  labels = NULL,
  limits = NULL,
  expand = c(0.01, 0),
  oob = scales::censor,
  na.value = NA_real_,
  trans = "identity",
  position = "left",
  ...
)

scale_x_comma(
  name = NULL,
  breaks = NULL,
  minor_breaks = NULL,
  labels = NULL,
  limits = NULL,
  expand = c(0.01, 0),
  oob = scales::censor,
  na.value = NA_real_,
  trans = "identity",
  position = "bottom",
  ...
)

scale_y_comma(
  name = NULL,
  breaks = NULL,
  minor_breaks = NULL,
  labels = NULL,
  limits = NULL,
  expand = c(0.01, 0),
  oob = scales::censor,
  na.value = NA_real_,
  trans = "identity",
  position = "left",
  ...
)
}
\arguments{
\item{name}{The name of the scale. Used as axis title.}

\item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item A numeric vector of positions
\item A function that takes the limits as input and returns breaks as output
}}

\item{minor_breaks}{One of:
\itemize{
\item \code{NULL} for no minor breaks
\item A numeric vector of positions
\item A function that given the limits returns a vector of minor breaks.
}}

\item{labels}{Specifying overrides the default format (i.e. you really don't want to do that). \code{NULL} means no labels.}

\item{limits}{A numeric vector of length two providing limits of the scale.
Use \code{NA} to refer to the existing minimum or maximum.}

\item{expand}{same as in ggplot2}

\item{oob}{Function that handles limits outside of the scale limits
(out of bounds). The default replaces out of bounds values with NA.}

\item{na.value}{If \code{na.translate = TRUE}, what value aesthetic
value should missing be displayed as?}

\item{trans}{Either the name of a transformation object, or the
object itself.}

\item{position}{The position of the axis. "left" or "right" for vertical
scales, "top" or "bottom" for horizontal scales}

\item{...}{passed on to \code{\link[scales:comma]{scales::comma_format()}} or \code{\link[scales:percent_format]{scales::percent_format()}}}
}
\description{
The \verb{_comma} ones set comma format for axis text and \code{expand=c(0,0)} (you need to set limits).
}
\details{
The \verb{_percent} ones set precent format for axis text and \code{expand=c(0,0)} (you need to set limits).
}
