% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_em.R
\name{theme_em}
\alias{theme_em}
\title{A \link{ggplot2} theme for personal use}
\usage{
theme_em(
  title_family = "MonoLisa emw",
  text_family = "MonoLisa emw",
  base_size = 10.5,
  text_color = "#2F2D2E",
  bg_color = "#fefefe",
  line_color = "gray50",
  plot_margin = ggplot2::margin(20, 20, 20, 20),
  md = FALSE
)
}
\arguments{
\item{title_family}{title elements font family}

\item{text_family}{text elements font family}

\item{base_size}{base font size}

\item{text_color}{color of the text & title elements}

\item{bg_color}{plot background color}

\item{line_color}{color of line elements}

\item{plot_margin}{plot margin specifications}

\item{md}{indicate whether to use markdown elements for text}
}
\description{
There is an option adapted from \code{theme_ib} to convert
theme use markdown theme elements (\code{md = TRUE}) to use markdown
theme elements from \code{ggtext} instead of \code{element_text()}.
}
