% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_em_dark.R
\name{theme_em_dark}
\alias{theme_em_dark}
\title{A \link{ggplot2} theme for dark personal use}
\usage{
theme_em_dark(
  title_family = "Architects Daughter",
  text_family = "MonoLisa emw",
  base_size = 10.5,
  title_color = "#ff6361",
  text_color = "#FFFDF7",
  bg_color = "#282a36",
  line_color = "#A0ACAD",
  plot_margin = ggplot2::margin(20, 20, 20, 20),
  with.grid = FALSE,
  with.line = FALSE,
  md = FALSE
)
}
\arguments{
\item{title_family}{title elements font family}

\item{text_family}{text elements font family}

\item{base_size}{base font size}

\item{title_color}{color of the title elements}

\item{text_color}{color of the text elements}

\item{bg_color}{plot background color}

\item{line_color}{color of line elements}

\item{plot_margin}{plot margin specifications}

\item{with.grid}{Include panel grid (default: FALSE)}

\item{with.line}{Include axis line (default: FALSE)}

\item{md}{indicate whether to use markdown elements for text}
}
\description{
There is an option adapted from \code{theme_ib} to convert
theme use markdown theme elements (\code{md = TRUE}) to use markdown
theme elements from \code{ggtext} instead of \code{element_text()}.
}
