% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_emw.R
\name{theme_emw}
\alias{theme_emw}
\title{Custom ggplot2 theme for Personal use}
\usage{
theme_emw(
  base_family = "Oswald",
  base_size = 10.5,
  with.panel.grid = FALSE,
  with.axis.line = TRUE,
  text.color = "#424651",
  title.color = "#424651",
  subtitle.color = "#6a6d76",
  plot.background.color = "#fefefe",
  axis.text.color = "#6a6d76",
  axis.text.size = base_size,
  subtitle.text.size = base_size * 1.35,
  title.text.size = base_size * 1.75,
  caption_family = base_family,
  caption_size = base_size * 3/4,
  caption_face = "plain",
  caption_margin = 12,
  title_margin = 10,
  subtitle_margin = 15,
  plot_margin = ggplot2::margin(20, 20, 20, 20),
  base_theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{base_family}{Base font family (default: "Metropolis")}

\item{base_size}{Base font size (default: 10.5)}

\item{with.panel.grid}{Include panel grid (default: FALSE)}

\item{with.axis.line}{Include axis line (default: TRUE)}

\item{text.color}{Text color (default: "#19647E")}

\item{title.color}{Title color (default: "#F8F8F2")}

\item{subtitle.color}{Subtitle color (default: "#6a6d76")}

\item{plot.background.color}{Plot background color (default: "white")}

\item{axis.text.color}{Axis text color (default: "#F8F8F2")}

\item{axis.text.size}{Axis text size (default: calculated)}

\item{subtitle.text.size}{Subtitle text size (default: calculated)}

\item{title.text.size}{Title text size (default: calculated)}

\item{caption_family}{Caption font family (default: base_family)}

\item{caption_size}{Caption font size (default: 10)}

\item{caption_face}{Caption font face (default: "plain")}

\item{caption_margin}{Caption margin (default: 12)}

\item{title_margin}{Title margin (default: 10)}

\item{subtitle_margin}{Subtitle margin (default: 15)}

\item{plot_margin}{Plot margin (default: specified)}

\item{base_theme}{Base ggplot2 theme (default: theme_minimal())}
}
\value{
A ggplot2 theme object.
}
\description{
Custom ggplot2 theme for Personal use
}
