% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_emw_dark.R
\name{theme_emw_dark}
\alias{theme_emw_dark}
\title{Custom dark ggplot2 theme for Personal use}
\usage{
theme_emw_dark(
  base_family = "Metropolis",
  base_size = 10.5,
  with.panel.grid = FALSE,
  with.axis.line = FALSE,
  text.color = "#e0e0e0",
  title.color = "#f1f1f1",
  subtitle.color = "#bcbcbe",
  plot.background.color = "#303036",
  axis.text.color = "#bcbcbe",
  axis.text.size = base_size,
  subtitle.text.size = base_size * 1.35,
  title.text.size = base_size * 1.75,
  title_margin = 10,
  subtitle_margin = 15,
  caption_family = base_family,
  caption_size = base_size * 3/4,
  caption_face = "plain",
  caption_margin = 12,
  plot_margin = ggplot2::margin(20, 20, 20, 20),
  base_theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{base_family}{Base font family (default: "Architects Daughter")}

\item{base_size}{Base font size (default: 10.5)}

\item{with.panel.grid}{Include panel grid (default: FALSE)}

\item{with.axis.line}{Include axis line (default: FALSE)}

\item{text.color}{Text color (default: "#f8f8f2")}

\item{title.color}{Title color (default: "#ff8281")}

\item{subtitle.color}{Title color (default: "#787b83")}

\item{plot.background.color}{Plot background color (default: "#424651")}

\item{axis.text.color}{Axis text color (default: "#f8f8f2",)}

\item{axis.text.size}{Axis text size (default: calculated)}

\item{subtitle.text.size}{Subtitle text size (default: calculated)}

\item{title.text.size}{Title text size (default: calculated)}

\item{title_margin}{Title margin (default: 10)}

\item{subtitle_margin}{Subtitle margin (default: 15)}

\item{caption_family}{Caption font family (default: base_family)}

\item{caption_size}{Caption font size (default: 10)}

\item{caption_face}{Caption font face (default: "plain")}

\item{caption_margin}{Caption margin (default: 14)}

\item{plot_margin}{Plot margin (default: specified)}

\item{base_theme}{Base ggplot2 theme (default: theme_minimal())}
}
\value{
A ggplot2 theme object.
}
\description{
Custom dark ggplot2 theme for Personal use
}
