% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_unl_dark.R
\name{theme_unl_dark}
\alias{theme_unl_dark}
\title{A \link{ggplot2} theme for dark personal use}
\usage{
theme_unl_dark(
  title_family = "Oswald",
  text_family = "Oswald",
  base_size = 10.5,
  title_color = "#D00000",
  subtitle_color = "#ffd74f",
  text_color = "#ebebea",
  bg_color = "#242423",
  line_color = "#c7c8ca",
  plot_margin = ggplot2::margin(20, 20, 20, 20),
  with.grid = FALSE,
  with.line = FALSE,
  md = FALSE
)
}
\arguments{
\item{title_family}{title elements font family}

\item{text_family}{text elements font family}

\item{base_size}{base font size}

\item{title_color}{color of the title elements}

\item{subtitle_color}{color of subtitle elements}

\item{text_color}{color of the text elements}

\item{bg_color}{plot background color}

\item{line_color}{color of line elements}

\item{plot_margin}{plot margin specifications}

\item{with.grid}{Include panel grid (default: FALSE)}

\item{with.line}{Include axis line (default: FALSE)}

\item{md}{indicate whether to use markdown elements for text}
}
\description{
There is an option to convert
theme use markdown theme elements (\code{md = TRUE}) to use markdown
theme elements from \code{ggtext} instead of \code{element_text()}.
}
