% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_colors.R
\name{unl_cols}
\alias{unl_cols}
\title{Function to extract unl brand colors as hex codes}
\usage{
unl_cols(...)
}
\arguments{
\item{...}{Character names of unl_colors}
}
\description{
Function to extract unl brand colors as hex codes
}
