% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_colors.R
\name{unl_get_colors}
\alias{unl_get_colors}
\title{Function to extract unl brand colors as hex codes}
\usage{
unl_get_colors(...)
}
\arguments{
\item{...}{Character names of unl_colors

Example: unl_get_colors("blue", "green")}
}
\value{
A vector of color hex codes
}
\description{
This function extracts University of Nebraska–Lincoln's brand colors in hex codes.
}
