% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_colors.R
\name{unl_pal}
\alias{unl_pal}
\title{A personal color palette to use with ggplot}
\usage{
unl_pal()
}
\description{
A personal color palette to use with ggplot
}
\examples{
library(scales)
scales::show_col(unl_pal()(7))
}
