% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_geoms.R
\name{update_geom_fonts_em}
\alias{update_geom_fonts_em}
\title{Update font defaults for text geoms to add handwriting text}
\usage{
update_geom_fonts_em(
  family = "Architects Daughter",
  face = "plain",
  size = 5,
  color = "#FF6361",
  hjust = 0
)
}
\arguments{
\item{family}{default geom font family}

\item{face}{default geom font face}

\item{size}{default geom font size}

\item{color}{default text color}

\item{hjust}{default hjust value}
}
\description{
Updates \link[ggplot2:geom_text]{ggplot2::geom_label}, \link[ggplot2:geom_text]{ggplot2::geom_text}, and \link[ggplot2:annotate]{ggplot2::annotate}
font defaults
}
