% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_emw.R
\name{update_geom_fonts_emw}
\alias{update_geom_fonts_emw}
\title{Update matching font defaults for text geoms}
\usage{
update_geom_fonts_emw(
  family = "Architects Daughter",
  face = "plain",
  size = 7,
  color = "#ff6361"
)
}
\arguments{
\item{family}{Font family, default "Architects Daughter".}

\item{face}{Font face, default "plain".}

\item{size}{font size in mm.}

\item{color}{Font color, default "#ff6361".}
}
\description{
Updates \link[ggplot2:geom_text]{ggplot2::geom_label} and \link[ggplot2:geom_text]{ggplot2::geom_text} font defaults
}
