% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_font.R
\name{check_font}
\alias{check_font}
\title{Check if a font is available}
\usage{
check_font(font_name)
}
\arguments{
\item{font_name}{The name of the font to check.}
}
\description{
Check if a font is available
}
