% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emw_palette.R, R/theme_emw.R
\name{emw_pal}
\alias{emw_pal}
\title{Custom dark color palette}
\usage{
emw_pal()

emw_pal()
}
\arguments{
\item{num_cols}{Number of colors to generate}

\item{var_type}{Variable type ("discrete" or "continuous")}
}
\value{
A vector of color codes.
}
\description{
Custom dark color palette

A bright qualitative color palette
}
\examples{
library(scales)
scales::show_col(emw_pal()(9))
}
