% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_emw.R
\name{emw_pal}
\alias{emw_pal}
\title{A personal color palette to use with ggplot}
\usage{
emw_pal()
}
\description{
A personal color palette to use with ggplot
}
\examples{
library(scales)
scales::show_col(emw_pal()(9))
}
