% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fonts.R
\name{load_pkg_fonts}
\alias{load_pkg_fonts}
\title{Load package fonts}
\usage{
load_pkg_fonts(verbose = TRUE)
}
\arguments{
\item{verbose}{Whether the newly registered fonts should be printed to the console. Defaults to \code{TRUE}.}
}
\description{
Registers fonts contained in {emwthemes} with \code{systemfonts::register_font()}. This function is called when the package is loaded.
}
\details{
All styles of a font family are registered as the Regular member of their own families in the form of "{family}-{style(s)}".

Call \code{systemfonts::registry_fonts()} after loading the fonts for list of fonts loaded.
Note also that displaying the font in \code{ggplot2} objects in rmd or quarto requires \code{showtext::showtext_auto()} in the code chunk. If saving with ``ggsave\code{run}showtext_opts(dpi = 300)`
}
\section{Fonts loaded}{

\describe{
\item{\code{Metropolis}}{ Find details later from \url{niceperson}. }
\item{\code{Oswald}}{ Google font, from \url{https://fonts.google.com/specimen/Oswald}. }
\item{\verb{Architects Daughter}}{ Google font, from \url{https://fonts.google.com/specimen/Architects+Daughter}. }
\item{\code{MonoLisaemw}}{ monospace font \url{findlater}. }
\item{\verb{Font Awesome 6 Free}}{ From \url{https://fontawesome.com/}.
The free version contains a subset of the Regular, Solid, and Brand styles:
\url{https://fontawesome.com/v6.15/icons?d=gallery&p=2&s=brands,regular,solid&m=free}.
}
}
}

