% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_em.R
\name{theme_em}
\alias{theme_em}
\title{A \link{ggplot2} theme for personal use}
\usage{
theme_em(
  title_family = "Permanent Marker",
  text_family = "Permanent Marker",
  base_size = 11,
  text_color = "#2F2D2E",
  bg_color = "#fefefe",
  line_color = "gray50",
  plot_margin = ggplot2::margin(0.5, 1, 0.5, 1, unit = "cm"),
  text_margin = ggplot2::margin(0.2, 0, 1, 0, unit = "cm"),
  md = FALSE
)
}
\arguments{
\item{title_family}{title elements font family}

\item{text_family}{text elements font family}

\item{base_size}{base font size}

\item{text_color}{color of the text & title elements}

\item{bg_color}{plot background color}

\item{line_color}{color of line elements}

\item{plot_margin}{plot margin specifications}

\item{text_margin}{margin around text}

\item{md}{indicate whether to use markdown elements for text}
}
\description{
There is an option adapted from \code{theme_ib} to convert
theme use markdown theme elements (\code{md = TRUE}) to use markdown
theme elements from \code{ggtext} instead of \code{element_text()}.
}
