% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_emw.R, R/theme_emw_dark.R
\name{theme_emw_dark}
\alias{theme_emw_dark}
\title{Custom dark ggplot2 theme for Personal use}
\usage{
theme_emw_dark(
  base_family = "daughter",
  base_size = 10.5,
  with.panel.grid = FALSE,
  text.color = "#ff6361",
  title.color = "#F8F8F2",
  plot.background.color = "#282a36",
  axis.text.color = "#f8f8f2",
  axis.text.size = base_size * 3/4,
  subtitle.text.size = base_size * 1.25,
  title.text.size = base_size * 1.75,
  caption_family = base_family,
  caption_size = 11,
  caption_face = "plain",
  caption_margin = 14,
  axis_title_just = "rt",
  plot_margin = ggplot2::margin(30, 30, 30, 30),
  base_theme = ggplot2::theme_minimal()
)

theme_emw_dark(
  base_family = "daughter",
  base_size = 10.5,
  with.panel.grid = FALSE,
  text.color = "#ff6361",
  title.color = "#F8F8F2",
  plot.background.color = "#282a36",
  axis.text.color = "#f8f8f2",
  axis.text.size = base_size * 3/4,
  subtitle.text.size = base_size * 1.25,
  title.text.size = base_size * 1.75,
  caption_family = base_family,
  caption_size = 11,
  caption_face = "plain",
  caption_margin = 14,
  axis_title_just = "rt",
  plot_margin = ggplot2::margin(30, 30, 30, 30),
  base_theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{base_family}{Base font family (default: "Architects Daughter")}

\item{base_size}{Base font size (default: 10.5)}

\item{with.panel.grid}{Include panel grid (default: FALSE)}

\item{text.color}{Text color (default: "#19647E")}

\item{title.color}{Title color (default: "#F8F8F2")}

\item{plot.background.color}{Plot background color (default: "#282a36")}

\item{axis.text.color}{Axis text color (default: "#f8f8f2")}

\item{axis.text.size}{Axis text size (default: calculated)}

\item{subtitle.text.size}{Subtitle text size (default: calculated)}

\item{title.text.size}{Title text size (default: calculated)}

\item{caption_family}{Caption font family (default: base_family)}

\item{caption_size}{Caption font size (default: 11)}

\item{caption_face}{Caption font face (default: "plain")}

\item{caption_margin}{Caption margin (default: 14)}

\item{axis_title_just}{Axis title justification (default: "rt")}

\item{plot_margin}{Plot margin (default: specified)}

\item{base_theme}{Base ggplot2 theme (default: theme_minimal())}
}
\value{
A ggplot2 theme object.

A ggplot2 theme object.
}
\description{
Custom dark ggplot2 theme for Personal use

Custom dark ggplot2 theme for Personal use
}
